Split boss up into pages, solve boss
